rpi/pyproject.toml
[project]
name = "sleep-quality-sensors"
version = "0.1.0"
description = "Raspberry Pi sensors agent for Sleep Quality Advisor"
requires-python = ">=3.14"
readme = "README.md"
license = { file = "../LICENSE" }
dependencies = [
  "bme680",
  "smbus2",
  "requests",
  "pydantic>=2",
  "python-dotenv>=1.0.0",
  "tenacity>=9.1.2",
  "sparkfun-qwiic-veml6030",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"

[tool.ruff.lint]
select = ["E","F","I","B","UP","SIM","I"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
mypy_path = "src"
plugins = []

[tool.uv]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-mock>=3.11.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
    "taskipy>=1.14.1",
    "types-requests",
]

[tool.taskipy.tasks]
lint = "ruff check ."
lint-fix = "ruff check . --fix"
format = "ruff format --check ."
format-fix = "ruff format ."
typecheck = "PYTHONPATH=src mypy --config-file mypy.ini src || PYTHONPATH=src mypy src"
check-fix = "task lint-fix && task format-fix && task typecheck && task test"
check = "task lint && task format && task typecheck && task test"
test = "PYTHONPATH=. pytest -q"
test-cov = "PYTHONPATH=. pytest -q --cov=src --cov-report=term-missing"
sensor-check = "PYTHONPATH=src python -m bme680_sensor.check_connection && PYTHONPATH=src python -m veml6030_sensor.check_connection"
rpi-run = "python -m src.main"


